always use the connectionMaker supplied remote address
The connectionMaker does bookkeeping based on the (string representation
of) the remote address. It is therefore important that we use that exact
same address in remoteConnection.remoteTCPAddr, since otherwise
connection lifecycle events such as termination will not find a
corresponding address in the connectionMaker's target map, and panic as
a result.

Fixes #59.