Allow multiple calls to ensureRecalculated() to wait on the same calculation
If another call is already waiting, wait on the same chan.

This is an improvement on the previous behaviour that only allowed one
ensureRecalculated() to proceed per calculate(), in the presence of
many recalculate() requests.

It is now possible for a later call to overtake the first.  Previously
it was possible for a later call to overtake an earlier call except
for the first.