Reduce the rate of route recalculation to reduce CPU load
Introduce a ticker to check twice a second if recalculation is needed.
This should help in situations where changes are hapenning very
rapidly so the recalculation was continuous.

Probably we should shut down the ticker when no requests are pending,
but this is a first step to see how it behaves.